requirements.txt
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
groq>=0.11.0
tavily-python>=0.5.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0